examples/maas-arena/requirements.txt
azure-ai-inference
azure-core
gradio
gradio-client
ipykernel
datasets
pandas